OutputForm객체 생성
출력형식을 변경하는 구조체를 만들었습니다.
