+++ LadderGame/LadderGame/Ladder.swift
@@ -12,8 +12,6 @@ import Foundation
 struct Ladder {
     
     var horizontalData = [[Bool]]()
-    var horizontalLadders = [[String]]()
-    
     
     mutating func makeLadderData(peopleNum: Int, ladderNum: Int) {
         let ladderWidth = peopleNum
@@ -23,17 +21,6 @@ struct Ladder {
         
         for dataIndex in 0 ..< ladderHeight {
             self.horizontalData[dataIndex] = ladderFloor.makeFloorRawData(rawData: self.horizontalData[dataIndex])
-        }
-    }
-    
-    mutating func makeHorizontalLadders(makeData: [[Bool]]) {
-        let ladderWidth = makeData[0].count
-        let ladderHeight = makeData.count
-        
-        self.horizontalLadders = Array(repeating: Array(repeating: "-", count: ladderWidth), count: ladderHeight)
-        
-        for index in 0 ..< ladderHeight {
-            self.horizontalLadders[index] = ladderFloor.makeFloorLadders(data: makeData[index], ladders: self.horizontalLadders[index])
         }
     }
 }
